Fix return type of fdo_blob_size
Previously, fdo_blob_size() was returning values of type int32_t, but
the return value was used as size_t in many places.

Since the typical return type of the size of a file is size_t, the
return type of fdo_blob_size() was changed to size_t.

This required following additional changes.
1. Update fdo_blob_size() to send 0 for error conditions.
2. The check on return value of fdo_blob_size() was updated to align
with this change.

+++ lib/credentials_from_file.c
@@ -196,7 +196,7 @@ bool read_normal_device_credentials(const char *dev_cred_file,
 				    fdo_dev_cred_t *our_dev_cred)
 {
 	bool ret = false;
-	int32_t dev_cred_len = 0;
+	size_t dev_cred_len = 0;
 	fdor_t *fdor = NULL;
 
 	if (!our_dev_cred) {
@@ -224,7 +224,7 @@ bool read_normal_device_credentials(const char *dev_cred_file,
 		return true;
 	}
 
-	LOG(LOG_DEBUG, "Reading DeviceCredential blob of length %"PRId32"\n", dev_cred_len);
+	LOG(LOG_DEBUG, "Reading DeviceCredential blob of length %"PRIu64"\n", dev_cred_len);
 
 	fdor = fdo_alloc(sizeof(fdor_t));
 	if (!fdor || !fdor_init(fdor) || !fdo_block_alloc_with_size(&fdor->b, dev_cred_len)) {
